+++ ignition-mobile/lib/core/constants.dart
@@ -1,0 +1,4 @@
+class AppConstants {
+  static const String appName = 'Ignition Pay';
+  static const String version = '0.1.0';
+}